beipack: add --build for PEP-517 projects
Add support for directly packing a PEP-517 project:

  beipack --build .

PEP-517 is a little bit too impactful on global state.  It depends on
the current working directory being set to the project root directory,
and writes output directly to stderr and stdout.  For this reason, we
wrap the invocation with the multiprocessing module.

Note: This is an extremely simplified (and therefore non-compliant)
build front-end without support for checking or installing dependencies.