beiboot: make --script an optional argument
This turns the old:

  beiboot script.py subcommand ...

into:

  beiboot --script script.py subcommand ...

which is roughly equivalent to:

  beiboot subcommand ... < script.py

And allows a new possibility:

  beiboot subcommand ...

which simply starts an interactive Python interpreter in a given
context, such as:

  beiboot ssh c

Additionally, add some smarts to recognise 'ssh' as a command name and
perform some adjustments to its arguments to account for the fact that
it expects a shell script, not an argument vector.  Also add 'container'
as a useful alias to start an interpreter via podman exec.

Finally, if run with no positional parameters, beiboot will just start
the interpreter in the local context.  This means that:

  beiboot

is roughly equivalent to:

  python -iq

But when piping a script:

  cat script.py | beiboot

avoids the repeated printing of >>> and ... lines.

The most interesting possibility introduced by this new interface is
that it's now possible to do things like execute beipacks created on the
fly:

  beipack ... | beiboot ssh c

Note: Somewhat annoyingly, the Python REPL appears to print an extra
newline on exit due to EOF, and I haven't been able to find a way to
disable that.  Scripts can avoid it by explicitly calling sys.exit().