new command: beipack-spawn
This allows creating a beipack which exists to run the specified command
with the specified set of files.

A trivial invocation looks like:

  beipack-spawn pip install cockpit-0-py3-none-any.whl

which will result in a beipack containing the wheel file.  When run, it
will create a temporary directory and extract the file into it, invoking
the given command.

By default, the last commandline argument is assumed to be the name of
the file, but this behaviour can be overridden by use of the --file
option which allows manually specifying one or more files to include.