beipack: use 'build' as our PEP-517 frontend
Remove our homebrew version of this and use the "full strength" official
version.

Among other things, this introduces support for in-tree build backends,
unbreaking beipack with Cockpit (which recently gained such a backend).

This also prevents us from accidentally depending on the non-stdlib
`toml` library.  The depend on `build` is soft: it is only required if
using `--build`.