beiboot: add --xz option
Use a two-stage bootloader approach to support decompression of .py.xz
files on the other side.

The bootloader communicates with us via ferny's InteractionAgent.  We
import ferny at function-level to avoid introducing a global dependency.

This is tested and working with at least ssh, sudo, and flatpak-spawn,
which are exactly the three things we plan to use it with for Cockpit.