test: don't exec() beipacks in tests
Running beipacks directly in the test environment is a bad idea: beipack
modifies the state of the import machinery and imports modules that
should only exist for one particular testcase.  There are ways to try to
reset the damage, but they're incomplete.

Let's avoid the problem in the first place: whenever we want to try to
run a beipack, let's do that in a separate process.