beipack: rework "template" logic
Remove support for custom templates — nobody is using that.

Rename 'beipack-template.py.template' to 'beipack_loader.py' and remove
some unused imports from the top (which are the only remaining linting
gripes).

Dynamically re-add the imports that we use from inside of beipack
itself.

Switch the entire structure of generating the beipack to be line-based
and paste it all together at the end.