Avoid using new importlib API before Python 3.11
The importlib.resources.abc stuff is pretty new: it doesn't exist before
3.11.  We're using that in our bootloader, which is the part that needs
to run everywhere, so clean that up some.

Our resources-collecting code is pretty much helpless without this
stuff, so disable the two test cases that depend on it unless we're
running on 3.11 or later.

This is enough to get things working on Python 3.9 and 3.10, so add them
to the CI workflow.