Drop setuptools depend to 59.6.0
This is the last version on PyPI which supports Python 3.6, which we
want to build against.

This version needs a setup.py file to work, but it seems like the basic
project settings can be read from pyproject.toml.  We still need a
setup.cfg for some of the file-finding and wrapper-script stuff.

This is enough to get Python 3.6 working, so add that to our CI.